replaced more appearance of double values by float values
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@7461 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/ConfigUpdate_p.java
@@ -257,7 +257,7 @@ public class ConfigUpdate_p {
         
         prop.putASIS("hintVersionDownload_versionResMain", (yacyVersion.latestMainRelease == null) ? "-" : yacyVersion.latestMainRelease.toAnchor());
         prop.putASIS("hintVersionDownload_versionResDev", (yacyVersion.latestDevRelease == null) ? "-" : yacyVersion.latestDevRelease.toAnchor());
-        prop.put("hintVersionAvailable_latestVersion", Double.toString(yacyVersion.latestRelease));
+        prop.put("hintVersionAvailable_latestVersion", Float.toString(yacyVersion.latestRelease));
          */
         
         return prop;
